22 09 26, ken 10481, 18강, 파라미터 타입 종류 설명, age를 Long 으로 받기

+++ src/main/java/com/ll/exam/app_2022_09_22/job/withParam/WithParamJobConfig.java
@@ -37,12 +37,11 @@ public class WithParamJobConfig {
                 .build();
     }
 
-
     @Bean
     @StepScope
     public Tasklet withParamStep1Tasklet(
             @Value("#{jobParameters['name']}") String name,
-            @Value("#{jobParameters['age']}") int age
+            @Value("#{jobParameters['age']}") Long age
     ) {
         return (contribution, chunkContext) -> {
             log.debug("name : {}, age : {}", name, age);
